third_party/WebKit/LayoutTests/http/tests/security/subresourceIntegrity/subresource-integrity-style-no-cors-no-xorigin.html

<!DOCTYPE html>
<head>
    <title>SRI on cross origin CORS disabled style, without CORS fetch</title>
    <style>.id1 { background-color: red }</style>
    <link rel="stylesheet"
          integrity="ni:///sha256;fy1lC6jbv7rOG6Flb00n5PYLggqe25IcE6QlZf8KtwI="
          href="http://localhost:8000/security/resources/cors-style.php?cors=false"></link>
</head>
<body>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        window.onload = function () {
            var ele = document.getElementById('id1');
            var divBackground = window.getComputedStyle(ele, null).getPropertyValue('background-color');
            assert_equals(divBackground, "rgb(255, 0, 0)");
            done();
        };
    </script>
    <div id="id1" class="id1"></div>
</body>
</html>
